requirements.txt
pdf2image==1.17.0
Pillow==10.4.0
pandas==2.2.3
openpyxl==3.1.5
